backend/advertisements/serializers.py
# backend/advertisements/serializers.py
from rest_framework import serializers
from .models import Advertisement

class AdvertisementSerializer(serializers.ModelSerializer):
    class Meta:
        model = Advertisement
        fields = ['image', 'alt_text', 'url']