backend/requirements.txt
django==4.2
djangorestframework==3.15
gunicorn==22.0
psycopg2-binary==2.9.9